+++ CHANGELOG.rst
@@ -7,8 +7,14 @@ Note that ``PHAB_ID=#`` and ``RB_ID=#`` correspond to associated messages in com
 Unreleased
 ----------
 
+New Features
+~~~~~~~~~~~~
+
+* finagle-core: Add Java compatibility for `c.t.f.Filter.TypeAgnostic.Identity`
+  via `c.t.f.Filter.typeAgnosticIdentity()`. ``PHAB_ID=D242006``
+
 * finagle-thrift: Add `c.t.finagle.thrift.MethodMetadata` which provides a `LocalContext` Key
-  for setting information about the current Thrift method and an accessor for retrieving 
+  for setting information about the current Thrift method and an accessor for retrieving
   the currently set value. ``PHAB_ID=D241295``
 
 18.11.0